+++ docs/기능 목록.md
@@ -69,7 +69,7 @@
 
 - [x] error 메시지 html로 이동
 - [x] form 요소에 submit 리스너 추가
-- [] 테스트코드 위치 변경
+- [x] 테스트 코드 위치 변경
 - [] AuthStore를 AuthChecker로 이동
 - [] 두 개의 다른 AuthChecker / 상속 이용
 - [] 핸들러 이름 변경
